+++ packages/maplibre-gl-export/src/lib/interfaces/ControlOptions.ts
@@ -21,7 +21,7 @@ export interface ControlOptions {
 	Crosshair?: boolean;
 	PrintableArea?: boolean;
 	Local?: Language;
-	AllowedSizes?: ('LETTER' | 'A2' | 'A3' | 'A4' | 'A5' | 'A6' | 'B2' | 'B3' | 'B4' | 'B5' | 'B6')[];
+	AllowedSizes?: ('Moonstrips' | 'Letter' | 'A2' | 'A3' | 'A4' | 'A5' | 'A6' | 'B2' | 'B3' | 'B4' | 'B5' | 'B6')[];
 	Filename?: string;
 	markerCirclePaint?: CirclePaint;
 	attributionOptions?: AttributionOptions;